more updates to tests
