--- apps/fund/tests/__init__.py
@@ -1,2 +1,0 @@
-from .unittests import *
-from .functional import *
